+++ tests/test_utils/test_system.py
@@ -6,6 +6,8 @@
 # ======================================================================================
 # flake8: noqa
 
+import sys
+
 import pytest
 
 from spectrochempy.utils.system import (
@@ -37,6 +39,10 @@ def test_is_kernel():
     assert not res
 
 
+@pytest.mark.skipif(
+    sys.platform == "win32",
+    reason="does not run well on windows (seems to be linked to some commit message)",
+)
 def test_sh():
     res = sh.git("show", "HEAD")
     assert res is not None