+++ docs/whatsnew/changelog.rst
@@ -49,6 +49,9 @@ Breaking changes
 ~~~~~~~~~~~~~~~~
 .. Add here new breaking changes (do not delete this comment)
 
+* Changed the default QP solver (quadprog -> osqp): The new solver is compatible with
+  python 3.11. Fastness and robusteness are improved. The quadprog solver can still be
+  used if available
 
 .. section